Prueba 7

+++ src/main/java/classroom/Asignatura.java
@@ -26,7 +26,7 @@ public class Asignatura {
 
     public Asignatura(String nombre, double codigoInterno, int codigoExterno,Tipo tipo) {
         this.nombre = nombre;
-        this.codigoInterno = codigoInterno;
+        this.codigoInterno = (int)codigoInterno;
         this.codigoExterno = codigoExterno;
         this.tipo = tipo;
     }
